requirements.txt
pandas
numpy
scikit-learn
nltk
beautifulsoup4
matplotlib
streamlit>=1.30.0
altair>=5.0.0
fpdf2
langgraph
python-dotenv
google-genai==1.70.0
joblib
sentence-transformers
groq
